Upgrade ktlint and disable no-unit-return rule (#1610)
* Rename to avoid confusions

This constant is used for the ktling gradle plugin but has no effect over ktlint version itself

* Update to latests version of the plugin

* Specify ktlint version explicitly

ktlint gradle plugin version does not target the latest ktlint release. override the plugin configuration by setting it explicitly

* Disable import-ordering

Has been promoted to be in standard ruleset in 0.34.0+ but it will have issues with intellij automatic ordering. Disable for now although in current discussion wether it should be there or not and might be removed in later versions

* Fix all curly-spacing violations

* Fix all parameter-list-wrapping violations

* Change printConsole(msg: String) from expression body to test no-unit-return

* Disable no-unit-return

Allow functions to declare Unit as its return type so impure functions can be made explicit